% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_diag_concept_doc.R
\docType{data}
\name{bc_diag_concept}
\alias{bc_diag_concept}
\title{Breast Cancer Diagnosis Concept Set.}
\format{
A data frame with \code{n} rows and the following columns:
\describe{
\item{code}{character. code string.}
\item{code_system}{character. vocabulary name (e.g., "icd10").}
\item{include}{logical. T for included codes, F for explicit exclusions.}
}
}
\source{
Created for the electronco package.
}
\description{
A data frame of diagnosis codes used to identify breast cancer in EHRs.
}
\examples{
head(bc_diag_concept)
}
