% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_proc_concept_doc.R
\docType{data}
\name{bc_proc_concept}
\alias{bc_proc_concept}
\title{Breast Cancer Procedures Concept Set.}
\format{
A data frame with 78 rows and the following columns:
\describe{
\item{code}{character. code string.}
\item{system}{character. vocabulary name (e.g., "cpt").}
\item{group}{character. procedure group.}
}
}
\source{
Created for the electronco package.
}
\description{
A data frame of procedure codes used with breast cancer in EHRs.
}
\examples{
head(bc_proc_concept)
}
