% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{.parse_to_date}
\alias{.parse_to_date}
\title{Extract diagnosis phenotype from diagnosis rows.}
\usage{
.parse_to_date(x, name = "date")
}
\arguments{
\item{diagnoses}{Dataset imported from a database or csv file.}

\item{concept}{Concept to look for.  For breast cancer, specify 'bc'.}

\item{lookback_start}{Beginning date of date range to look at, in the format YYYY-MM-DD (year-month-day).}

\item{lookback_end}{End data of date range to look at, in the format YYYY-MM-DD (year-month-day).}

\item{min_events}{The minimum number of occurences of a given concept.}

\item{patient_id_col}{Name of the patient_id column in the "diagnoses" dataset.}

\item{code_col}{Name of the code column in the "diagnoses" dataset.}

\item{date_col}{Name of the diagnosis_date column in the "diagnoses" dataset.}

\item{system_col}{Name of the code_type column in the "diagnoses" dataset.}
}
\description{
Inputs:
\itemize{
\item diagnoses: tibble with the following columns: patient_id, code, code_system, date (Date).
\item concept_set: tibble with the following columns: code, code_system, include (logical TRUE=include, FALSE=exclude).
}
}
\details{
Output: list(patient_level, evidence, metadata).
}
