% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.R
\name{pheno}
\alias{pheno}
\title{Produce Composite Phenotype Flags}
\usage{
pheno(
  spec,
  meds_tbl = NULL,
  proc_tbl = NULL,
  labs_tbl = NULL,
  dx_tbl = NULL,
  return_evidence = TRUE
)
}
\arguments{
\item{spec}{A list of id, version, date_range, rule, and min_count.}

\item{meds_tbl}{The output of meds() or a vector of medications to filter for.}

\item{proc_tbl}{The output of proc() or a vector of procedures to filter for.}

\item{labs_tbl}{The output of labs() or a vector of labs to filter for.}

\item{dx_tbl}{The output of diagnosis() or a vector of diagnoses to look for.
If using the output of diagnosis, be sure to specify the patients level list
(e.g. dx_tbl = diagnosis_output$patient_level).}

\item{return_evidence}{Logical; if TRUE include evidence_sample list-column in the result.}
}
\value{
A tibble with columns: patient_id, flag, flag_date, evidence_count, evidence_sample, spec_id, spec_version.
}
\description{
Produce Composite Phenotype Flags
}
