% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{diagnosis}
\alias{diagnosis}
\title{Extract diagnosis phenotype from diagnosis rows (dplyr style)}
\usage{
diagnosis(
  diagnoses,
  concept_set,
  params = list(lookback_start = as.Date("1900-01-01"), lookback_end = Sys.Date(),
    min_occurrences = 1, ),
  patient_id_col = "patient_id",
  code_col = "code",
  system_col = "code_system",
  date_col = "date",
  source_col = "source",
  encounter_col = "encounter_id"
)
}
\description{
Inputs:
\itemize{
\item diagnoses: tibble with the following columns: patient_id, code, code_system, date (Date).
\item concept_set: tibble with the following columns: code, code_system, include (logical TRUE=include, FALSE=exclude).
\item params: list(lookback_start, lookback_end, min_occurrences)
}
}
\details{
Output: list(patient_level, evidence, metadata)
}
