% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{diagnosis}
\alias{diagnosis}
\title{Extract diagnosis phenotype from diagnosis rows.}
\usage{
diagnosis(
  diagnoses,
  concept_set,
  params = list(lookback_start = as.Date("1900-01-01"), lookback_end = Sys.Date(),
    min_occurrences = 1, ),
  patient_id_col = "patient_id",
  code_col = "code",
  system_col = "code_type",
  date_col = "diagnosis_date"
)
}
\arguments{
\item{diagnoses}{Dataset imported from a database or csv file.}

\item{concept_set}{Path to concept_set.}

\item{params}{List of parameters supplied by the user (lookback_start, lookback_end, and min_occurences)}

\item{patient_id_col}{Name of the patient_id column in the "diagnoses" dataset.}

\item{code_col}{Name of the code column in the "diagnoses" dataset.}

\item{system_col}{Name of the code_type column in the "diagnoses" dataset.}

\item{date_col}{Name of the diagnosis_date column in the "diagnoses" dataset.}
}
\description{
Inputs:
\itemize{
\item diagnoses: tibble with the following columns: patient_id, code, code_system, date (Date).
\item concept_set: tibble with the following columns: code, code_system, include (logical TRUE=include, FALSE=exclude).
\item params: list(lookback_start, lookback_end, min_occurrences)
}
}
\details{
Output: list(patient_level, evidence, metadata).
}
