% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{diagnosis}
\alias{diagnosis}
\title{Extract diagnosis phenotype from diagnosis data.}
\usage{
diagnosis(
  diagnoses,
  concept,
  lookback_start,
  lookback_end,
  min_events = 1,
  patient_id_col = "patient_id",
  code_col = "code",
  system = "code_type",
  date_col = "diagnosis_date",
  date_format = "\%Y-\%m-\%d"
)
}
\arguments{
\item{diagnoses}{Dataset imported from a database or csv file.}

\item{concept}{Concept to look for. For breast cancer, specify 'bc'.}

\item{lookback_start}{Beginning date of date range to look at in YYYY-MM-DD (year-month-day) format.}

\item{lookback_end}{End date of date range to look at in YYYY-MM-DD (year-month-day) format.}

\item{min_events}{The minimum number of occurrences of a given concept.}

\item{patient_id_col}{Name of the patient_id column in the "diagnoses" dataset.}

\item{code_col}{Name of the code column in the "diagnoses" dataset.}

\item{system}{Name of the code_type column in the "diagnoses" dataset.}

\item{date_col}{Name of the diagnosis_date column in the "diagnoses" dataset.}

\item{date_format}{R date format of the diagnosis_date column in the "diagnoses" dataset (e.g. '\%Y-\%m-\%d').}
}
\description{
Inputs:
\itemize{
\item diagnoses: tibble or data frame with the following columns: patient_id, code, code_type, diagnosis_date.
\item concept_set (included with package): tibble with the following columns: code, code_type, include (logical TRUE=include, FALSE=exclude).
}
}
\details{
Output: list(patient_level, evidence, metadata).
}
