% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labs.R
\name{labs}
\alias{labs}
\title{Extract lab marker cohorts.}
\usage{
labs(
  data,
  markers,
  match_type = c("exact", "contains"),
  patient_id_col = "patient_id",
  lab_date_col = "lab_date",
  lab_name_col = "lab_name",
  date_range = NULL,
  cohort_type = c("any", "first", "last", "all"),
  min_tests = 1
)
}
\arguments{
\item{data}{Dataframe or tibble with input data.}

\item{markers}{Vector of the markers to look for.}

\item{match_type}{How to match the values specified in "markers".
Can be 'exact' or 'contains'.}

\item{patient_id_col}{Name of the patient_id column in data.}

\item{lab_date_col}{Name of the lab_date column in data.}

\item{lab_name_col}{Name of the lab_name column in data.}

\item{date_range}{Beginning and ending dates to look for.}

\item{cohort_type}{Type of cohort to find.
Can be 'any', 'first', 'last', or 'all'.}

\item{min_tests}{Minimum number of occurrences per patient to filter for.}
}
\value{
a tibble whose shape depends on cohort_type
}
\description{
Extract lab marker cohorts.
}
